requirements.txt
streamlit
pypdf
transformers
torch
sentence-transformers
scikit-learn
faiss-cpu
pandas